requirements.txt
Flask
pyswisseph
pytz
requests
gunicorn
Flask-Cors
tqdm
python-dotenv
rich